requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
az-iranian-bank-gateways==2.1.1
binaryornot==0.4.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
crispy-bootstrap4==2024.10
Django==5.2.3
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-tailwind==4.0.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
lxml==6.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
phonenumbers==8.13.54
phonenumberslite==9.0.12
pillow==11.1.0
platformdirs==4.4.0
pycryptodome==3.23.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
urllib3==2.5.0
zeep==4.3.1
